+++ src/redux/types.ts
@@ -1,0 +1,6 @@
+export interface UserCredentials {
+    [index: string]: any;
+    displayName: string,
+    email: string,
+    password: string,
+};
